@@ -12,7 +12,7 @@
         }
     ],
     "require": {
-        "php": ">=5.4"
+        "php": ">=7.1"
         },
     "autoload": {
         "psr-4": {
